#3 Moved skill selector to a separate component
