+++ frontend/src/components/ItemTypeDefinitions.tsx
@@ -1,5 +1,5 @@
 import { ItemType } from '../types/ItemTypes';
-import ItemRender from './dragdrop/ItemRender';
+import ItemRender from './skills/ItemRender';
 
 export const ItemTypes: Record<string, ItemType> = {
     SKILL: {
