+++ frontend/src/components/skills/ItemRender.tsx
