+++ frontend/src/components/skills/SelectedSkill.tsx
