+++ frontend/src/components/skills/Skill.tsx
